requirements.txt
Flask==1.1.2
werkzeug==1.0.1
jinja2==2.11.3
markupsafe==2.0.1
itsdangerous==1.1.0
psycopg2-binary==2.9.1
